+++ platform/src/codatplatform/models/operations/delete_api_key.py
@@ -20,9 +20,12 @@ class DeleteAPIKeyRequest:
 @dataclasses.dataclass
 class DeleteAPIKeyResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""