+++ platform/src/codatplatform/models/operations/get_webhook.py
@@ -21,10 +21,13 @@ class GetWebhookRequest:
 @dataclasses.dataclass
 class GetWebhookResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""
     webhook: Optional[shared_webhook.Webhook] = dataclasses.field(default=None)
     r"""OK"""